Finalize groundwork for the Blackfin port
git-svn-id: svn+ssh://svn.gna.org/svn/xenomai/trunk@218 c6d672ea-8702-0410-b560-f74c916a59fe

+++ include/asm-bfinnommu/wrappers.h
@@ -26,6 +26,6 @@
 
 #include <asm-generic/xenomai/wrappers.h> /* Read the generic portion. */
 
-#define wrap_access_ok(task,addr,size) _access_ok((addr),(size))
+#define wrap_access_ok(task,addr,size) _access_ok((unsigned long)(addr),(size))
 
 #endif /* _XENO_ASM_BFINNOMMU_WRAPPERS_H */
